Commenting/Cleanup/Bugfixing

+++ PlatformerEngine/inc/levelgenerator.h
@@ -5,7 +5,7 @@
 #include "player.h"
 #include "global.h"
 
-u8 getTileValue(s16 x, s16 y);
+u16 getTileValue(s16 x, s16 y);
 
 void freeCollisionMap();
 void generateCollisionMap(const u8 map[][48]);
